requirements.txt
Pillow
requests
urllib3
easyocr